# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lounge Queue API Documentation",
        "description": "API documentation for the Lounge Queue system",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/examination": {
            "get": {
                "tags": [
                    "Examination"
                ],
                "summary": "Get current visitor's examination",
                "description": "Retrieves the current examination details for the authenticated visitor. Requires visitor authentication.",
                "operationId": "getExaminationByCurrentVisitor",
                "responses": {
                    "200": {
                        "description": "Examination details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "examination_id": {
                                                    "description": "Unique identifier of the examination",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "provider_id": {
                                                    "description": "ID of the provider conducting the examination",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "provider_name": {
                                                    "description": "Full name of the provider",
                                                    "type": "string",
                                                    "example": "Dr. Smith"
                                                },
                                                "started_at": {
                                                    "description": "When the examination started",
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-03-29T10:00:00Z"
                                                },
                                                "status": {
                                                    "description": "Current status of the examination",
                                                    "type": "string",
                                                    "enum": [
                                                        "in_progress",
                                                        "completed",
                                                        "cancelled"
                                                    ],
                                                    "example": "in_progress"
                                                },
                                                "duration": {
                                                    "description": "How long the examination has been running",
                                                    "type": "string",
                                                    "example": "5 minutes"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No active examination found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No active examination found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing authentication token"
                    },
                    "403": {
                        "description": "Forbidden - User is not a visitor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/provider/examination": {
            "get": {
                "tags": [
                    "Examination"
                ],
                "summary": "Get current provider's examination",
                "description": "Retrieves the current examination details for the authenticated provider. Requires provider authentication.",
                "operationId": "getExaminationByCurrentProvider",
                "responses": {
                    "200": {
                        "description": "Examination details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "examination_id": {
                                                    "description": "Unique identifier of the examination",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "visitor_id": {
                                                    "description": "ID of the visitor being examined",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "visitor_name": {
                                                    "description": "Full name of the visitor",
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "started_at": {
                                                    "description": "When the examination started",
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-03-29T10:00:00Z"
                                                },
                                                "status": {
                                                    "description": "Current status of the examination",
                                                    "type": "string",
                                                    "enum": [
                                                        "in_progress",
                                                        "completed",
                                                        "cancelled"
                                                    ],
                                                    "example": "in_progress"
                                                },
                                                "duration": {
                                                    "description": "How long the examination has been running",
                                                    "type": "string",
                                                    "example": "5 minutes"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No active examination found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No active examination found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing authentication token"
                    },
                    "403": {
                        "description": "Forbidden - User is not a provider"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}